[ot] hw/opentitan: simplify trace-events message types
* ibex_get_current_pc() returns 32-bit values
* `io_read_out` and `io_write` use 32-bit value for pc and addr

+++ hw/opentitan/ot_aon_timer.c
@@ -1,7 +1,7 @@
 /*
  * QEMU OpenTitan AON Timer device
  *
- * Copyright (c) 2023 Rivos, Inc.
+ * Copyright (c) 2023-2024 Rivos, Inc.
  *
  * Author(s):
  *  Loïc Lefort <loic@rivosinc.com>
@@ -332,8 +332,8 @@ static uint64_t ot_aon_timer_read(void *opaque, hwaddr addr, unsigned size)
         break;
     }
 
-    uint64_t pc = ibex_get_current_pc();
-    trace_ot_aon_timer_read_out((unsigned)addr, REG_NAME(reg), val32, pc);
+    uint32_t pc = ibex_get_current_pc();
+    trace_ot_aon_timer_read_out((uint32_t)addr, REG_NAME(reg), val32, pc);
 
     return (uint64_t)val32;
 }
@@ -347,8 +347,8 @@ static void ot_aon_timer_write(void *opaque, hwaddr addr, uint64_t value,
 
     hwaddr reg = R32_OFF(addr);
 
-    uint64_t pc = ibex_get_current_pc();
-    trace_ot_aon_timer_write((unsigned)addr, REG_NAME(reg), val32, pc);
+    uint32_t pc = ibex_get_current_pc();
+    trace_ot_aon_timer_write((uint32_t)addr, REG_NAME(reg), val32, pc);
 
     switch (reg) {
     case R_ALERT_TEST:
